Last commit for performers
The paper is extremely unclear to be honest and i have tried to implement this for a couple of days but to no avail. The training works fine but the loss, t/step and results are horrible. There must be a flaw with my implementation but this is likely my last attempt on this for a while.